fix transfer api to cast bigint to string
